requirements.txt
python-dotenv>=1.0.0
langgraph>=1.1.0
langchain-openai>=1.2.0
langchain-community>=0.4.0
langchain-huggingface>=1.2.0
langchain-text-splitters>=1.1.0
watchdog>=6.0.0
llama-parse>=0.6.0
